RavenDB-20310 Indexes view: toggle is not changing the text & positions of elements
