+++ src/Raven.Studio/typescript/components/common/Type.scss
@@ -10,7 +10,7 @@ a.link {
     cursor: pointer;
 }
 
-a:not(.btn):not(.no-decor) {
+a:not(.btn):not(.dropdown-item):not(.no-decor) {
     background-image: linear-gradient(90deg, $border-color-light, $border-color-light);
     background-position: 50% 100%;
     background-repeat: no-repeat;
